+++ src/Neo.Compiler.CSharp/MethodConvert/ConstructorConvert.cs
@@ -12,6 +12,7 @@ extern alias scfx;
 
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Neo.IO;
 using Neo.VM;
 using System.Linq;
