+++ src/Neo.Compiler.CSharp/MethodConvert/PropertyConvert.cs
@@ -12,6 +12,7 @@ extern alias scfx;
 
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Neo.IO;
 using Neo.SmartContract;
 using Neo.SmartContract.Native;
 using Neo.VM;
