+++ tests/Neo.Compiler.CSharp.UnitTests/DebugAndTestBase.cs
@@ -20,7 +20,7 @@ public class DebugAndTestBase<T> : TestBase<T>
         TestSingleContractBasicBlockStartEnd(context!);
     }
 
-    private static void TestSingleContractBasicBlockStartEnd(CompilationContext result)
+    public static void TestSingleContractBasicBlockStartEnd(CompilationContext result)
     {
         try
         {
@@ -30,54 +30,55 @@ public class DebugAndTestBase<T> : TestBase<T>
             }
             catch
             {
-                Console.WriteLine($"Omitted: {result.ContractName}");
+                Console.WriteLine($"Omited: {result.ContractName}");
             }
         }
         catch
         {
             Console.WriteLine($"Error compiling: {result.ContractName}");
+            return;
         }
     }
 
-    private static void TestSingleContractBasicBlockStartEnd(NefFile nef, ContractManifest manifest, JToken debugInfo)
+    public static void TestSingleContractBasicBlockStartEnd(NefFile nef, ContractManifest manifest, JToken debugInfo)
     {
         var basicBlocks = new ContractInBasicBlocks(nef, manifest, debugInfo);
 
         // TODO: support CALLA and do not return
 
-        var instructions = basicBlocks.GetScriptInstructions().ToList();
-        var (_, _, jumpTargets) =
+        List<VM.Instruction> instructions = basicBlocks.GetScriptInstructions().ToList();
+        (_, _, Dictionary<VM.Instruction, HashSet<VM.Instruction>> jumpTargets) =
             Neo.Optimizer.JumpTarget.FindAllJumpAndTrySourceToTargets(instructions);
 
         Dictionary<VM.Instruction, VM.Instruction> nextAddrTable = new();
         VM.Instruction? prev = null;
-        foreach (var i in instructions)
+        foreach (VM.Instruction i in instructions)
         {
             if (prev != null)
                 nextAddrTable[prev] = i;
             prev = i;
         }
 
-        foreach (var basicBlock in basicBlocks.basicBlocks)
+        foreach (BasicBlock basicBlock in basicBlocks.basicBlocks)
         {
             // Basic block ends with allowed OpCodes only, or the next instruction is a jump target
             Assert.IsTrue(OpCodeTypes.allowedBasicBlockEnds.Contains(basicBlock.instructions.Last().OpCode) || jumpTargets.ContainsKey(nextAddrTable[basicBlock.instructions.Last()]));
             // Instructions except the first are not jump targets
-            foreach (var i in basicBlock.instructions.Skip(1))
+            foreach (VM.Instruction i in basicBlock.instructions.Skip(1))
                 Assert.IsFalse(jumpTargets.ContainsKey(i));
             // Other instructions in the basic block are not those in allowedBasicBlockEnds
-            foreach (var i in basicBlock.instructions.Take(basicBlock.instructions.Count - 1))
+            foreach (VM.Instruction i in basicBlock.instructions.Take(basicBlock.instructions.Count - 1))
                 Assert.IsFalse(OpCodeTypes.allowedBasicBlockEnds.Contains(i.OpCode));
         }
 
         // Each jump target starts a new basic block
-        foreach (var target in jumpTargets.Keys)
+        foreach (VM.Instruction target in jumpTargets.Keys)
             Assert.IsTrue(basicBlocks.basicBlocksByStartInstruction.ContainsKey(target));
 
         // Each instruction is included in only 1 basic block
-        var includedInstructions = new HashSet<VM.Instruction>();
-        foreach (var basicBlock in basicBlocks.basicBlocks)
-            foreach (var instruction in basicBlock.instructions)
+        HashSet<VM.Instruction> includedInstructions = new();
+        foreach (BasicBlock basicBlock in basicBlocks.basicBlocks)
+            foreach (VM.Instruction instruction in basicBlock.instructions)
             {
                 Assert.IsFalse(includedInstructions.Contains(instruction));
                 includedInstructions.Add(instruction);
